Tests: improvements, small fixes, and re-organization of files  #50
- Reorganized conversation repository test in smaller files
- Improved test for new_conversation
- Randomized research context and conversation picked (now it's not just the first one)
